requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
black==23.9.1
cachetools==5.3.1
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
decorator==5.1.1
distlib==0.3.7
executing==1.2.0
filelock==3.12.3
flake8==6.1.0
frozenlist==1.4.0
idna==3.4
ipython==8.15.0
jedi==0.19.0
matplotlib-inline==0.1.6
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
pluggy==1.3.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.0
pyflakes==3.1.0
Pygments==2.16.1
pyproject-api==1.6.1
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
stack-data==0.6.2
tox==4.11.3
traitlets==5.9.0
urllib3==2.0.4
virtualenv==20.24.5
wcwidth==0.2.6
yarl==1.9.2
